Update multi-channel-cell-counter.ijm
Added the ability to have no thresholding with "None"
Added the ability to change whether Watershed is used and incorporated the Adjustable Watershed plugin
Added the ability to get cell intensities for all channels (even the counter-stain. To get stats for the RAW counter-stain images, non-thresholded, just add a duplicate channel. The first one will be taken as the counter-stain).